+++ src/sfft.rs
@@ -1,5 +1,5 @@
 use num_complex::Complex;
-use std::f32::consts::PI;
+use std::f64::consts::PI;
 
 fn reverse_bits(n: u64, bits: u64) -> u64 {
     let mut result = 0;
@@ -10,7 +10,7 @@ fn reverse_bits(n: u64, bits: u64) -> u64 {
     result
 }
 
-fn reorder(complex_data: &mut Vec<Complex<f32>>, log2n: u64) {
+fn reorder(complex_data: &mut Vec<Complex<f64>>, log2n: u64) {
     for i in 0..complex_data.len() as u64 {
         let x = reverse_bits(i, log2n);
         if x > i {
@@ -19,7 +19,7 @@ fn reorder(complex_data: &mut Vec<Complex<f32>>, log2n: u64) {
     }
 }
 
-fn cooley_tukey(complex_data: &mut Vec<Complex<f32>>, inv: bool) {
+fn cooley_tukey(complex_data: &mut Vec<Complex<f64>>, inv: bool) {
     let sign = if inv { 1.0 } else { -1.0 };
     let mut log2n: u64 = 0;
     let n = complex_data.len();
@@ -32,8 +32,8 @@ fn cooley_tukey(complex_data: &mut Vec<Complex<f32>>, inv: bool) {
     for i in 1..=log2n {
         let m = 1 << i;
         let deltawn = Complex {
-            re: f32::cos(2.0 * PI / m as f32),
-            im: sign * f32::sin(2.0 * PI / m as f32),
+            re: f64::cos(2.0 * PI / m as f64),
+            im: sign * f64::sin(2.0 * PI / m as f64),
         };
 
         for k in (0..complex_data.len()).step_by(m) {
@@ -50,12 +50,12 @@ fn cooley_tukey(complex_data: &mut Vec<Complex<f32>>, inv: bool) {
     }
 }
 
-pub fn fft(complex_data: &mut Vec<Complex<f32>>) {
+pub fn fft(complex_data: &mut Vec<Complex<f64>>) {
     cooley_tukey(complex_data, false);
 }
 
-pub fn ifft(complex_data: &mut Vec<Complex<f32>>) {
+pub fn ifft(complex_data: &mut Vec<Complex<f64>>) {
     cooley_tukey(complex_data, true);
-    let n = complex_data.len() as f32;
+    let n = complex_data.len() as f64;
     complex_data.iter_mut().for_each(|x| *x = *x / n);
 }
